Add files via upload
When using Classical shadow for properties prediction. We always not reconstruct the full state via shadows because it takes an exponential amount of resources. One better method is to use Median-of-Means estimation to predict fidelities between the system and the target state. Here I add median-of-means estimator function, to effeciently predict n-qubit system properties.